@@ -1,8 +1,8 @@
 export const environment = {
   production: true,
-  httpLinkServer: 'http://10.42.0.1:5000',
-  wsLinkServer: 'ws://10.42.0.1:5000',
-  downloadLinkServer: 'http://10.42.0.1:5000/download',
+  httpLinkServer: 'http://10.42.0.1:6000',
+  wsLinkServer: 'ws://10.42.0.1:6000',
+  downloadLinkServer: 'http://10.42.0.1:6000/download',
   envName: 'Production',
   appName: 'GONDOLIN'
 };
