@@ -10,11 +10,14 @@ export const UserQuery = {
 
   async me(parent: any, args: any, ctx: Context, info: GraphQLResolveInfo) {
     const id = getUserId(ctx);    
-    console.log(id);
     return ctx.db.query.user({where: {id: id } }, info);
   },
 
   async user(parent: any, args: any, ctx: Context, info: GraphQLResolveInfo) {
     return ctx.db.query.user({where: { id: args.id }}, info);
+  },
+
+  async usersConnection(parent: any, args: any, ctx: Context, info: GraphQLResolveInfo) {
+    return ctx.db.query.usersConnection({...args}, info);
   }
 }
